// packages/frontend/generated/api/schemas/notificationDigestGroup.ts
/**
 * Generated by orval v7.13.0 🍺
 * Do not edit manually.
 * Chase Light API
 * GitHub リポジトリ監視サービス API - TypeScript + Hono + Zod + OpenAPI
 * OpenAPI spec version: 1.0.0
 */
import type { NotificationDigestGroupActivityType } from './notificationDigestGroupActivityType';
import type { NotificationDigestEntry } from './notificationDigestEntry';

export interface NotificationDigestGroup {
  /** アクティビティ種別 */
  activityType: NotificationDigestGroupActivityType;
  entries: NotificationDigestEntry[];
}