// packages/frontend/generated/api/schemas/notificationDigestGroupActivityType.ts
/**
 * Generated by orval v7.13.0 🍺
 * Do not edit manually.
 * Chase Light API
 * GitHub リポジトリ監視サービス API - TypeScript + Hono + Zod + OpenAPI
 * OpenAPI spec version: 1.0.0
 */

/**
 * アクティビティ種別
 */
export type NotificationDigestGroupActivityType = typeof NotificationDigestGroupActivityType[keyof typeof NotificationDigestGroupActivityType];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const NotificationDigestGroupActivityType = {
  issue: 'issue',
  pull_request: 'pull_request',
  release: 'release',
} as const;