// packages/frontend/generated/api/schemas/notificationListItem.ts
/**
 * Generated by orval v7.13.0 🍺
 * Do not edit manually.
 * Chase Light API
 * GitHub リポジトリ監視サービス API - TypeScript + Hono + Zod + OpenAPI
 * OpenAPI spec version: 1.0.0
 */
import type { NotificationSummary } from './notificationSummary';
import type { NotificationDataSource } from './notificationDataSource';

export interface NotificationListItem {
  notification: NotificationSummary;
  dataSources: NotificationDataSource[];
}