// packages/frontend/generated/api/schemas/notificationListResponseData.ts
/**
 * Generated by orval v7.13.0 🍺
 * Do not edit manually.
 * Chase Light API
 * GitHub リポジトリ監視サービス API - TypeScript + Hono + Zod + OpenAPI
 * OpenAPI spec version: 1.0.0
 */
import type { NotificationListItem } from './notificationListItem';
import type { NotificationListPageInfo } from './notificationListPageInfo';

export type NotificationListResponseData = {
  items: NotificationListItem[];
  pageInfo: NotificationListPageInfo;
};